Show slightly better diffs, still needs work
This is one of the challenges of not using external deps. The options
are to use what Node has built-in (which is what this is) or create and
include a diff function. Unfortunately what's built-in isn't great, but
it's better than simply printing X !== y. For now I'll leave it like
this. Diffing isn't that hard, but at a point the no-dep model
encounters diminishing returns when you start implementing all external
functionality within the single module.