All writes now go through reporters, more to do
This ensures all writing to console is done via reporter. However, as
of now the reporter is little more than a proxy for console writing.
Instead it should process and collect data for the various reporting
mechanisms. For now that's just the standard interface, will be adding
tap as well.